[chore] cleanup, Events are now just examples and not actually used

+++ ext_tables.sql
@@ -34,5 +34,5 @@ CREATE TABLE `tx_notifications_framework_domain_model_notification`
 
 CREATE TABLE fe_users
 (
-	`notifications` tinyint(8) unsigned DEFAULT '0' NOT NULL
+	`notifications` int(11) unsigned DEFAULT '0' NOT NULL
 );
